KVM: PPC: Handle ultracalls when secure memory facility is disabled
When the Secure Memory Facility (SMF) is disabled (SMFCTRL[E]=0), the
hardware will send system calls with LEV=2 (ultracalls) to the
hypervisor. The ultracall opcode space is guaranteed to not clash with
any hypercalls.

That means that in a nested scenario we could have KVM behaving as a
fake Ultravisor for the nested hypervisor/nested guest, which is
interesting for development and testing because we can stress the
secure guest code without having SMF support in hardware or firmware.

This patch adds the initial code for directing the ultracalls initiated by
the nested hypervisor and nested guest to KVM in L0.

For the nested guest (L2), hypercalls are currently handled by its
nested hypervisor (L1) after the H_ENTER_NESTED call returns. From now
on, before returning from the H_ENTER_NESTED, we first check if L2
issued an ultracall and handle that in L0 instead.

For the nested hypervisor (L1), the path is the same as the current
one so we first go through kvmppc_pseries_do_hcall to handle any
hypercalls and fallback to kvmppc_pseries_do_ucall.

---
I have opted for 'ucall' functions instead of reusing the 'hcall' ones
to keep the naming consistent with the ultravisor API.