KVM: PPC: Book3S HV: Add worker thread for ultracalls
The ultravisor API describes a few hypercalls that can be issued by
the firmware to the hypervisor in behalf of the secure guest. These
hypercalls can be issued (several times) during the handling of a
single ultracall. The hypervisor sees them as if they were issued by
the guest.

To mimic this behavior when Secure Memory Facility is disabled and KVM
is acting as an Ultravisor in a nested scenario, we need to take into
consideration that ultracalls issued by the nested guest (L2) will
reach KVM L0 during the H_ENTER_NESTED call from L1.

So to allow L1 to handle any subsequent hcall done by L0 as if they
came from L2, we need to return from H_ENTER_NESTED with the
appropriate registers set up as if L2 was issuing the hypercall.

After L1 handles the call, H_ENTER_NESTED will be called again to
enter L2, at which point L0 has the opportunity to look at the
hypercall results before actually going back into L2. This creates a
particular flow:

 L2 -------> L0 HV ----------------> L1 HV ----------------> L0 HV -------------> L2
     ucall            return from           H_ENTER_NESTED          return from
                    H_ENTER_NESTED                                     ucall
               |______________________ xN _____________________|

This patch implements a worker thread for ultracalls that need to be
handled in such a way, so we can stop the worker thread when it needs
to wait for L1 to handle a hypercall and continue after L1 has called
H_ENTER_NESTED again. The worker function will return U_TOO_HARD if it
needs to wait for L1 and any other valid return code when it wants to
go back to L2.

---
I have a included a dummy implementation for UV_ESM to facilitate the
review.