KVM: PPC: Book3S HV: Add worker thread for ultracalls
The ultravisor API describes a few hypercalls that can be issued by
the ultravisor firmware to the hypervisor in behalf of the secure
guest (SVM). These hypercalls can be issued (several times) during the
handling of an ultracall. Therefore it is possible that the following
execution flow happens:

  SVM -------> UV -------> HV -------> UV -------> SVM
       ucall       hcall       ucall       urfid
               |__________ xN _________|

When Secure Memory Facility is disabled and KVM is acting as an
Ultravisor in a nested scenario, ultracalls issued by the nested guest
(L2) will reach KVM L0 during the H_ENTER_NESTED call from L1. Like
so:

  L1 ----------------> L0 -------------> L2 -------> L0
      H_ENTER_NESTED       enter guest       ucall

So to allow L1 to handle any subsequent hcall done by L0, we need to
first return from H_ENTER_NESTED.

The nested hypervisor (L1) will then handle the hypercall as if it
came from L2 and it will enter the nested guest again:

  L0 ----------------> L1 ---------------> L1 ----------------> L0 -->
       return from         handles hcall       H_ENTER_NESTED
      H_ENTER_NESTED

.. and L0 can then continue with the ultracall handling and ultimately
return to the nested guest with the ultracall result:

  --> L0 ---------------> L0 -------------> L2
          handles ucall       return from
                                ucall

Adapting the first diagram to the nested scenario, it would look like:

 L2 -------> L0 HV ----------------> L1 HV ----------------> L0 HV -------------> L2
     ucall            return from           H_ENTER_NESTED          return from
                    H_ENTER_NESTED                                     ucall
               |______________________ xN _____________________|

This patch implements a worker thread for ultracalls that need to be
handled in such a way, so we can stop the worker thread when it needs
to wait for L1 to handle a hypercall and continue after L1 has called
H_ENTER_NESTED again. The worker function will return U_TOO_HARD if it
needs to wait for L1 and any other valid return code when it wants to
go back to L2.

---
I have a included prototypes for UV_ESM to facilitate the review.