KVM: PPC: Book3S HV: Implement UV_REGISTER_MEM_SLOT
The UV_REGISTER_MEM_SLOT ultracall is used by the hypervisor to inform
the ultravisor about the memory ranges of its guest that will soon
become secure.

In a nested environment, when KVM L0 is acting as an ultravisor for
the nested guest L2, the ultracall behaves the same and we need to
have a way of saving the memslot information.

This patch (re)implements a basic infrastructure for keeping track of
memslots, given that i) the generic KVM memslot code is not exposed to
the architecture and ii) the nested guest doesn't have a 'struct kvm'
that we can use in L0 to keep the memslots.

The implementation is on purpose kept similar to the generic code in
case we choose to address the aforementioned issues in the future.

The memslot's rmap will be used to keep track of ultravisor-specific
information about the guest frames such as the paging and sharing
state. The dirty_bitmap, userspace_addr and flags fields are for now
unused.