KVM: PPC: Book3S HV: Don't clear the shadow_pgtable when emulating an ultravisor
As part of H_TLB_INVALIDATE we currently clear the shadow partition
table in kvmhv_flush_nested:

L0:
 kvmhv_flush_nested
 kvmhv_emulate_tlbie_lpid
 kvmhv_do_nested_tlbie
 kvmppc_pseries_do_hcall
 kvmppc_vcpu_run_hv

When emulating an ultravisor, we don't want the nested hypervisor to
be able to clear the whole shadow partition table, but only the shared
entries that are shared with it by L0.

The sharing of pages is not yet implemented so for now not clearing
the shadow partition table during H_TBL_INVALIDATE is sufficient.