KVM: PPC: Book3S HV: Handle paging ultracalls when SMF is disabled
The ultravisor firmware has an isolated region of memory where it
copies pages to/from when the hypervisor issues UV_PAGE_IN/OUT calls.

When the Secure Memory Facility is not present in a machine and we are
using KVM to handle the ultracalls there is no special memory region
where to move the pages. However, in the nested scenario the L0 host
can still stop the nested hypervisor from accessing pages in its own
memory by manipulation of that L1 guest's partition scoped tables.

This patch implements the page tables manipulation as follows:

UV_PAGE_IN:
    Unmap the gra -> hra pte from L0 partition scoped tables for L1
    so that L1 cannot directly access L2's memory.

    Make sure there is a nested gra -> hra translation in L0's shadow
    partition scoped tables for L2.

UV_PAGE_OUT:
    Reinsert the gra -> hra translation in L0 partition scoped tables
    for L1.