KVM: PPC: Book3S HV: Reserve L2 memory for RTAS bounce buffers
With secure memory facility (SMF) enabled, the buffer used by the
secure guest during RTAS calls cannot be shared with the hypervisor
because sharing is done in a page granularity and the RTAS buffer size
might not be a multiple of the page size. Sharing a page in this
scenario would leave unencrypted data exposed in the buffer.

Although this feature is meant to be run with SMF disabled, i.e. when
there is no data protection concerns, it is more useful if we keep the
overall flow similar to what a real ultravisor would do.

This patch implements the reservation in the nested guest's memory of
a few pages to be used instead of the regular RTAS buffer during RTAS
calls.