powerpc/pseries/svm: Check kernel command line for svm=on
The KVM code is being modified to support emulating an ultravisor so
that we are able to run the secure guest code in a machine without the
Secure Memory Facility. In that environment the MSR[S] bit will never
be set so we need some other way of knowing that the guest made the
secure switch.

For now let's check the kernel command line for the svm=on option
since at the point is_secure_guest is called, we will only be running
if the secure switch was successful.