KVM: PPC: Book3S HV: Share rtas bounce buffers
When KVM is emulating an ultravisor, RTAS calls are done through a set
of pages divided into bounce buffers so that no content of the nested
guest's memory is shared with the nested hypervisor aside from the
RTAS Argument Call Buffer and any other buffers pointed to by
individual arguments.

This patch implements the above by sharing the RTAS bounce buffer
pages with L1 during UV_ESM and splitting them into smaller chunks for
use during each RTAS call.

In L2 the RTAS bounce buffer pages are marked as reserved in the
device-tree (previous patch), but there is a need to ensure that the
nested hypervisor has not invalidated the translations by the time the
RTAS call happens. We "pin" the RTAS pages by sharing them under a
lock that is also protecting the invalidation process that stems from
UV_PAGE_INVAL.